requirements.txt
pillow==11.3.0
psutil==7.0.0
smbus2==0.5.0
wiringpi==2.32.1
pillow==11.3.0
psutil==7.0.0
smbus2==0.5.0
wiringpi==2.32.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
pillow==11.3.0
psutil==7.0.0
requests==2.32.4
smbus2==0.5.0
urllib3==2.5.0
wiringpi==2.32.1
